Complete roulette system overhaul with working ball positioning
- Rewrote RouletteController with BagerTest's exact working logic
- Added RouletteTestManager for comprehensive testing and verification
- Added RouletteSetupHelper for easy one-click setup
- Improved BagerTest with enhanced debugging
- Fixed ball positioning issues - now works 100% correctly
- Added realistic ball animation during spin
- Moved all scripts to correct directory structure

Features:
- Space: Spin roulette with input number
- B: Test BagerTest directly
- R: Reset positions
- 1,2,3,4: Quick test specific numbers
- Auto-verification of results
- Detailed debug logging
